- moved `createGuessedCountriesMap` to helpers - added tests for it
